Move Spoiler class
