+++ src/main/java/org/owasp/wrongsecrets/challenges/kubernetes/Challenge7.java
@@ -2,7 +2,7 @@ package org.owasp.wrongsecrets.challenges.kubernetes;
 
 
 import org.owasp.wrongsecrets.ScoreCard;
-import org.owasp.wrongsecrets.Spoiler;
+import org.owasp.wrongsecrets.challenges.Spoiler;
 import org.owasp.wrongsecrets.challenges.Challenge;
 import org.owasp.wrongsecrets.challenges.ChallengeEnvironment;
 import org.owasp.wrongsecrets.challenges.ChallengeNumber;
